requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
install==1.3.5
mysqlclient==2.2.1
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
